ML/requirements.txt
# =============================================================================
# AIOps ML — Dépendances Python
# =============================================================================
# Installer avec : pip3 install -r requirements.txt
# =============================================================================

# Machine Learning
scikit-learn==1.5.2
prophet==1.1.6

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Prometheus API
prometheus-api-client==0.5.5

# Configuration (.env)
python-dotenv==1.0.1
